# documentation/api/swagger.yaml
openapi: 3.0.0
info:
  title: PulsePlus Game Platform API
  version: 1.0.0
  description: API documentation for the PulsePlus gaming and engagement platform
servers:
  - url: /api
    description: Main API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    CurrencyBalance:
      type: object
      properties:
        coins:
          type: integer
        gems:
          type: integer
        points:
          type: integer
    CurrencyTransfer:
      type: object
      properties:
        to_competitor_id:
          type: string
          format: uuid
        amount:
          type: integer
          minimum: 1
        currency_type:
          type: string
          enum: [COINS, GEMS, POINTS]
        reason:
          type: string
    TransactionHistory:
      type: object
      properties:
        sys_id:
          type: string
          format: uuid
        competitor_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [TRANSFER, PURCHASE, REWARD, REFUND]
        amount:
          type: integer
        currency_type:
          type: string
          enum: [COINS, GEMS, POINTS]
        balance_after:
          type: integer
        reference_id:
          type: string
          format: uuid
        reference_type:
          type: string
        created_at:
          type: string
          format: date-time
    Shop:
      type: object
      properties:
        sys_id:
          type: string
          format: uuid
        game_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        icon_url:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        currency_type:
          type: string
          enum: [COINS, GEMS, POINTS]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
paths:
...
  /economy/balance/{competitorId}:
    get:
      tags:
        - Economy
      summary: Get competitor's currency balance
      security:
        - BearerAuth: []
      parameters:
        - name: competitorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyBalance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /economy/transfer:
    post:
      tags:
        - Economy
      summary: Transfer currency between competitors
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyTransfer'
      responses:
        '200':
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /economy/history/{competitorId}:
    get:
      tags:
        - Economy
      summary: Get competitor's transaction history
      security:
        - BearerAuth: []
      parameters:
        - name: competitorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionHistory'
                  total:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /economy/shops:
    post:
      tags:
        - Economy
      summary: Create a new shop
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shop'
      responses:
        '201':
          description: Shop created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /economy/shops/{shopId}:
    get:
      tags:
        - Economy
      summary: Get shop details
      security:
        - BearerAuth: []
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Shop details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
security:
  - BearerAuth: [] 